Vinyls Oh My - A Music sampler Website
created a site to view 5 music samples from 4 different musical genres.

there's a landing page as well 4 sections. within the 4 sections, contains an image and a "jukebox" sub section with 5 audio files and the text "Your browser does not support the audio element." just in case the file can't run on a browser.